+++ Magic/CHANGELOG.md
@@ -36,6 +36,7 @@
  - The `/mwarp go` and `/send` commands will try to load the target world if it is not loaded
  - Add a third alt-cast spell to wands, for a total of 4 possible fixed spells
  - Add SuperWalls spells to default configs
+ - Allow numbers (0 - 1) in reset_mana_on_activate wand property, to reset to a fixed number
 
 # 7.11